refactor bookmarks into BookmarkChangeProcessor
