+++ components/brave_sync/tools.cc
@@ -1,8 +1,11 @@
 /* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
+#include "brave/components/brave_sync/tools.h"
+
 #include <string>
 #include <sstream>
+
 #include "crypto/random.h"
 #include "base/strings/string_util.h"
 #include "base/time/time.h"
